backend/requirements.txt
django>=2,<2.1
django-cors-headers>=2.2.0,<2.3
django-environ>=0.4.4,<0.5
djangorestframework>=3.8,<4
djangorestframework-gis>=0.12,<1.0
djangorestframework-jwt>=1.11.0,<2
djoser>=1.1.5,<1.2
django-rest-swagger>=2.1.2,<2.2
psycopg2-binary>=2.7,<3
flake8
